# [CPU] Faster reduce kernel for SHM allreduce