This PR make reduce kernel in SHM based allreduce faster by using `#pragma omp parallel for`.

On a server with 2 socket and SNC4 (total 8 ranks), the allreduce time can reduce from 15us to 12us for 10KB message size, and 100us to 60us for 100KB message size.